backend/requirements.txt
fastapi
uvicorn[standard]
dnspython
pydantic
python-dotenv
